Impl lbi_from_int, redesign interface & library
Instead of focusing on managing a pool of bigints, I've redesigned the
library to focus on managing memory on single bigints. This makes the
implementation simple. The speed of implementation matters at this time.

While there's a trick to be able to support syntax like the following:
lbi_drop(foo, bar, baz);
I've elected to keep it simple and just implement functions that act on
a single object. It aint hard to write N-1 more lines of code, and it is
more easier to reason about than macro magic.

TODO: I will have to rewrite the examples again... *sigh*
TODO: while reading magnitude_foo is pleasant, typing it isnt. Refactor
it to make it shorter.